fix data types

+++ services/rust/common/src/dimension.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct Dimension {
-    pub id: i32,
+    pub id: i16,
     pub name: String
 }
